refactor: TableGroup, OrderTable 도메인을 table 패키지로 분리
