+++ src/main/java/kitchenpos/table/domain/OrderTableRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.repository;
+package kitchenpos.table.domain;
 
 import java.util.List;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
