+++ src/main/java/kitchenpos/table/dto/request/OrderTableNumberOfGuestsRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.request;
+package kitchenpos.table.dto.request;
 
 public class OrderTableNumberOfGuestsRequest {
     private int numberOfGuests;
