+++ src/main/java/kitchenpos/table/dto/response/TableGroupResponse.java
@@ -1,9 +1,8 @@
-package kitchenpos.dto.response;
+package kitchenpos.table.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.table.domain.TableGroup;
 
 public class TableGroupResponse {
     private Long id;
