+++ src/test/java/kitchenpos/table/ui/TableGroupRestControllerTest.java
@@ -1,26 +1,13 @@
-package kitchenpos.ui;
+package kitchenpos.table.ui;
 
-import static kitchenpos.ui.MenuGroupRestControllerTest.메뉴_그룹_생성_요청;
-import static kitchenpos.ui.MenuRestControllerTest.메뉴_생성_요청;
-import static kitchenpos.ui.OrderRestControllerTest.주문_생성_요청;
-import static kitchenpos.ui.ProductRestControllerTest.상품_생성_요청;
-import static kitchenpos.ui.TableRestControllerTest.좌석_생성_요청;
+import static kitchenpos.table.ui.TableRestControllerTest.좌석_생성_요청;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Arrays;
-import java.util.List;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.Product;
-import kitchenpos.domain.TableGroup;
-import kitchenpos.dto.request.MenuProductRequest;
-import kitchenpos.dto.request.MenuRequest;
-import kitchenpos.dto.request.TableGroupRequest;
-import kitchenpos.dto.response.OrderTableResponse;
-import kitchenpos.dto.response.TableGroupResponse;
+import kitchenpos.BaseTest;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.dto.request.TableGroupRequest;
+import kitchenpos.table.dto.response.TableGroupResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.core.ParameterizedTypeReference;
@@ -28,7 +15,7 @@ import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
-class TableGroupRestControllerTest extends BaseTest {
+public class TableGroupRestControllerTest extends BaseTest {
     private OrderTable 좌석_요청_1;
     private OrderTable 좌석_요청_2;
     private TableGroupRequest 좌석_그룹_요청;
